+++ modules/permissions/lib/policy-delegate-factory.js
@@ -3,6 +3,7 @@
 var assert = require('assert');
 var GroupPolicyDelegate = require('./policies/group-policy-delegate');
 const GlGroupPolicyDelegate = require('./policies/gl-group-policy-delegate');
+const GlProjectPolicyDelegate = require('./policies/gl-project-policy-delegate');
 var GhRepoPolicyDelegate = require('./policies/gh-repo-policy-delegate');
 var GhOrgPolicyDelegate = require('./policies/gh-org-policy-delegate');
 var GhUserPolicyDelegate = require('./policies/gh-user-policy-delegate');
@@ -18,6 +19,9 @@ function policyDelegateFactory(userId, userLoader, securityDescriptor) {
 
     case 'GL_GROUP':
       return new GlGroupPolicyDelegate(userId, userLoader, securityDescriptor);
+
+    case 'GL_PROJECT':
+      return new GlProjectPolicyDelegate(userId, userLoader, securityDescriptor);
 
     case 'GH_REPO':
       return new GhRepoPolicyDelegate(userId, userLoader, securityDescriptor);