+++ modules/permissions/test/policies/gl-project-policy-delegate-test.js
@@ -3,89 +3,89 @@
 const assert = require('assert');
 const proxyquireNoCallThru = require('proxyquire').noCallThru();
 
-describe('gl-group-policy-delegate', () => {
-  const GROUP1_USER = { _id: 1, username: 'x_gitlab' };
-  const NOT_GROUP1_USER = { _id: 2, username: 'y_gitlab' };
+describe('gl-project-policy-delegate', () => {
+  const PROJECT1_USER = { _id: 1, username: 'x_gitlab' };
+  const NOT_PROJECT1_USER = { _id: 2, username: 'y_gitlab' };
   const NON_GITLAB_USER = { _id: 2, username: 'y' };
   const INVALID_USER = { _id: 3 };
   const USER_API_THROWS_ERROR = { _id: 666 };
-  const GROUP1 = 'group1';
+  const PROJECT1 = 'group/project1';
 
   const FIXTURES = [
     {
-      name: 'is group member',
-      group: GROUP1,
-      user: GROUP1_USER,
-      policy: 'GL_GROUP_MEMBER',
+      name: 'is project member',
+      project: PROJECT1,
+      user: PROJECT1_USER,
+      policy: 'GL_PROJECT_MEMBER',
       expectedResult: true
     },
     {
-      name: 'is not group member',
-      group: GROUP1,
-      user: NOT_GROUP1_USER,
-      policy: 'GL_GROUP_MEMBER',
+      name: 'is not project member',
+      project: PROJECT1,
+      user: NOT_PROJECT1_USER,
+      policy: 'GL_PROJECT_MEMBER',
       expectedResult: false
     },
     {
-      name: 'is group maintainer',
-      group: GROUP1,
-      user: GROUP1_USER,
-      policy: 'GL_GROUP_MAINTAINER',
+      name: 'is project maintainer',
+      project: PROJECT1,
+      user: PROJECT1_USER,
+      policy: 'GL_PROJECT_MAINTAINER',
       expectedResult: true
     },
     {
-      name: 'is not group maintainer',
-      group: GROUP1,
-      user: NOT_GROUP1_USER,
-      policy: 'GL_GROUP_MAINTAINER',
+      name: 'is not project maintainer',
+      project: PROJECT1,
+      user: NOT_PROJECT1_USER,
+      policy: 'GL_PROJECT_MAINTAINER',
       expectedResult: false
     },
     {
       name: 'is not a GitLab user',
-      group: GROUP1,
+      project: PROJECT1,
       user: NON_GITLAB_USER,
-      policy: 'GL_GROUP_MEMBER',
+      policy: 'GL_PROJECT_MEMBER',
       expectedResult: false
     },
     {
       name: 'anonymous',
-      group: GROUP1,
+      project: PROJECT1,
       user: null,
-      policy: 'GL_GROUP_MEMBER',
+      policy: 'GL_PROJECT_MEMBER',
       expectedResult: false
     },
     {
       name: 'user sans username',
-      group: GROUP1,
+      project: PROJECT1,
       user: INVALID_USER,
-      policy: 'GL_GROUP_MEMBER',
+      policy: 'GL_PROJECT_MEMBER',
       expectedResult: false
     },
     {
       name: 'invalid policy',
-      group: GROUP1,
+      project: PROJECT1,
       user: INVALID_USER,
       policy: 'INVALID',
       expectedResult: false
     },
     {
       name: 'API throws error',
-      group: GROUP1,
+      project: PROJECT1,
       user: USER_API_THROWS_ERROR,
       policy: 'INVALID',
       expectedResult: 'throw'
     }
   ];
 
-  let GlGroupPolicyDelegate;
-  function StubGitLabGroupService(user) {
+  let GlProjectPolicyDelegate;
+  function StubGitLabProjectService(user) {
     this.getMembership = async (uri, username) => {
       if (user === USER_API_THROWS_ERROR) {
-        throw new Error('My fake StubGitLabGroupService error from USER_API_THROWS_ERROR');
+        throw new Error('My fake StubGitLabProjectService error from USER_API_THROWS_ERROR');
       }
 
-      if (uri === GROUP1) {
-        if (user === GROUP1_USER && username === user.username) {
+      if (uri === PROJECT1) {
+        if (user === PROJECT1_USER && username === user.username) {
           return {
             accessLevel: 50,
             isMember: true,
@@ -105,7 +105,7 @@ describe('gl-group-policy-delegate', () => {
   }
 
   function stubGetIdentityForUser(user, provider) {
-    if ((provider === 'gitlab' && user === GROUP1_USER) || user === NOT_GROUP1_USER) {
+    if ((provider === 'gitlab' && user === PROJECT1_USER) || user === NOT_PROJECT1_USER) {
       return { fakeProvider: true, providerKey: user.username };
     }
 
@@ -113,21 +113,24 @@ describe('gl-group-policy-delegate', () => {
   }
 
   before(() => {
-    GlGroupPolicyDelegate = proxyquireNoCallThru('../../lib/policies/gl-group-policy-delegate', {
-      'gitter-web-gitlab': {
-        GitLabGroupService: StubGitLabGroupService
-      },
-      'gitter-web-identity': {
-        getIdentityForUser: stubGetIdentityForUser,
-        GITLAB_IDENTITY_PROVIDER: 'gitlab'
+    GlProjectPolicyDelegate = proxyquireNoCallThru(
+      '../../lib/policies/gl-project-policy-delegate',
+      {
+        'gitter-web-gitlab': {
+          GitLabProjectService: StubGitLabProjectService
+        },
+        'gitter-web-identity': {
+          getIdentityForUser: stubGetIdentityForUser,
+          GITLAB_IDENTITY_PROVIDER: 'gitlab'
+        }
       }
-    });
+    );
   });
 
   FIXTURES.forEach(meta => {
     it(meta.name, async () => {
       const securityDescriptor = {
-        linkPath: meta.group
+        linkPath: meta.project
       };
 
       const user = meta.user;
@@ -137,7 +140,7 @@ describe('gl-group-policy-delegate', () => {
         return Promise.resolve(user);
       }
 
-      const delegate = new GlGroupPolicyDelegate(userId, userLoader, securityDescriptor);
+      const delegate = new GlProjectPolicyDelegate(userId, userLoader, securityDescriptor);
 
       try {
         const result = await delegate.hasPolicy(meta.policy);